Fix: The Delete / Get Message Query Conditions
Fixed issue 1: if a user deleted his message the user still can
get his deleted message because it only deleted from server only if
the sender and the recipient both deleted the message.

Fixed issue 2: the recipient of the message could not delete
the message, but he did delete the message for the sender.